% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df}
\alias{df}
\title{A Simulation dataset containing 500 cells and 80 genes}
\format{
A dataset with 500 rows and 82 cols
\describe{
\item{Index}{index of cells}
\item{Time}{normalized pseudotime of each cell}
\item{Gene1}{expression counts of gene1}
...
}
}
\usage{
data("df")
}
\description{
A Simulation dataset containing 500 cells and 80 genes
}
\keyword{datasets}
