% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotimeEstInfer.R
\name{estimation}
\alias{estimation}
\title{Using PSO algorithm to estimate parameters within the fitting distribution}
\usage{
estimation(y, t, marginal, iter = 50)
}
\arguments{
\item{iter}{}
}
\value{

}
\description{
Using PSO algorithm to estimate parameters within the fitting distribution
}
