% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotimeAPI.R
\name{Fisher_info}
\alias{Fisher_info}
\title{Compute Fisher information matrix of interested parameters}
\usage{
Fisher_info(t, para, marginal)
}
\arguments{
\item{t}{A numeric vector of the input normalized pseodotime data of a given gene,
length equals the numbers of cells}

\item{para}{A numeric vector of the Estimated parameters of the predefined distribution,
length equals 6 or 7 depending on distribution(take 6 for Poisson or ZIP, 7 for NB and ZINB)}

\item{marginal}{A string of the distribution name. One of \code{Poisson}, \code{ZIP}, \code{NB} and \code{ZINB}.}
}
\value{
A 4x4 Fisher information matrix.
}
\description{
Compute Fisher information matrix of interested parameters
}
\examples{
para<-c(0.1,0.2,0.3,0.4,0.5,0.6)
t<-runif(100, min = 0, max = 1)
marginal<-"ZIP"
Fisher_info(t=t, para=para,marginal=marginal)

}
