% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotimeEstInfer.R
\name{inference}
\alias{inference}
\title{Compute estimated Fisher information matrix and covariance matrix}
\usage{
inference(t, para, marginal)
}
\arguments{
\item{marginal}{}
}
\value{

}
\description{
Compute estimated Fisher information matrix and covariance matrix
}
