% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_result.R
\name{plot_result}
\alias{plot_result}
\title{Plot the Fitted Model}
\usage{
plot_result(
  para,
  t,
  color,
  marginal,
  flag,
  y1,
  gene_name = NULL,
  save_dir = NULL
)
}
\arguments{
\item{para}{A numeric vector of the Estimated parameters of the predefined distribution,
length equals 6 or 7 depending on distribution(take 6 for Poisson or ZIP, 7 for NB and ZINB)}

\item{t}{A numeric vector of the input normalized pseudotime data of a given gene,
length equals the numbers of cells}

\item{color}{A string vector of length 4 to define plot color, default=\code{c('red', 'blue', 'orange', 'darkgreen')}}

\item{marginal}{A string of the distribution name. One of \code{Poisson}, \code{ZIP}, \code{NB} and \code{ZINB}.}

\item{flag}{A boolean variable, flag=T indicates Valley shape, flag=F indicates Hill shape}

\item{y1}{A vector of integers, representing the input expression counts of a given gene,
length equals the numbers of cells}

\item{gene_name}{A vector of strings, indicates the genes' name used in the model, shown in plotting,
default=NULL}

\item{save_dir}{A vector of strings, indicates saving path of plots, default=NULL(does not save)}
}
\value{
A ggplot object
}
\description{
Plot the Fitted Model
}
\examples{
data("df")
t<-df$Time
marginal<-"ZIP"
color<-c('red', 'blue', 'orange', 'darkgreen')

#Case1
flag<-FALSE
para<-c(2.29,3.27,11.79,0.58,30.4,60.82)
y1<-df$Gene1
plot_result(para, t, color, marginal, flag, y1, "Gene1","~/Desktop/Jessica_lab/scGTM_result/")

#Case2
flag<-TRUE
para<-c(2.96143,3.769441,2.098308,0.4638821,2.971249,-2.451574)
y1<-df$Gene11
plot_result(para, t, color, marginal, flag, y1, "Gene11","~/Desktop/Jessica_lab/scGTM_result/")


}
\author{
Shiyu Ma, Lehan Zou
}
