% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotimeAPI.R
\name{pso_obj_fct}
\alias{pso_obj_fct}
\title{Compute log_likelihood cost function of one gene based on given parameters}
\usage{
pso_obj_fct(b, y, t, marginal)
}
\arguments{
\item{b}{A numeric vector of the parameters of the predefined distribution,
length equals 6 or 7 depending on distribution(take 6 for Poisson or ZIP, 7 for NB and ZINB)}

\item{y}{A integer vector of the input expression counts data of a given gene,
length equals the numbers of cells}

\item{t}{A numeric vector of the input normalized pseodotime data of a given gene,
length equals the numbers of cells}

\item{marginal}{A string of the distribution name. One of \code{Poisson}, \code{ZIP}, \code{NB} and \code{ZINB}.}
}
\value{
The log_likelihood cost of the gene
}
\description{
The log_likelihood cost of fitting a gene's expression counts and pseudotime into a
self-defined marginal distribution and parameters
}
\examples{
b<-c(0.1,0.2,0.3,0.4,0.5,0.6)
y<-floor(runif(100, min = 0, max = 20))
t<-runif(100, min = 0, max = 1)
marginal<-"ZIP"
pso_obj_fct(b,y=y, t=t, marginal=marginal)

}
\author{
Shiyu Ma
}
