% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runscGTM.R
\name{runscGTM}
\alias{runscGTM}
\title{Estimate Parameters in Single-cell Gene Expression Generalized Trend Model on a List of Genes}
\usage{
runscGTM(
  t,
  y,
  sce = NULL,
  marginal = "ZIP",
  iter_num = 50,
  seed = 123,
  mc.cores = 2
)
}
\arguments{
\item{t}{A numeric vector of the input normalized pseudotime data of a given gene,
length equals the numbers of cells(If sce is not null, t is a string of gene names to use in the model)}

\item{y}{A tibble, representing the input expression counts of corresponding lists of genes,
number of rows equals the numbers of cells, number of columns equals the numbers of targeted genes,
(If sce is not null, y is a SingleCellExperiment object with counts data)}

\item{sce}{A character vector, indicates the assay name applied to the SingleCellExperiment object
default=NULL}

\item{marginal}{A string of the distribution name. One of \code{Poisson}, \code{ZIP}, \code{NB}, \code{ZINB} and \code{Gaussian}.
default=\code{ZIP}}

\item{iter_num}{A single integer vector, indicates max number of iteration used in the PSO algorithm
that estimates model parameters}

\item{seed}{A numeric variable of the random seed, affecting parametric fitting of the marginal distribution.
default=123}

\item{mc.cores}{Number of cores used for computing.}
}
\value{
A tibble of summary results of genes
}
\description{
Estimate Parameters in Single-cell Gene Expression Generalized Trend Model on a List of Genes
}
\examples{
data("df")
res <- scGTM::runscGTM(t=df$Time, y=df[,3:5])

data("sce")
t_sce<-rownames(sce)
res_sce <- runscGTM(t=t_sce, y=sce, sce="logcounts", marginal="Gaussian")

}
\author{
Shiyu Ma
}
