% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sce}
\alias{sce}
\title{A SingelCellExperiment dataset containing 2,000 genes and 2,291 cells}
\format{
A dataset with 2,000 rows and 2,291 cols
}
\usage{
data("sce")
}
\description{
A SingelCellExperiment dataset containing 2,000 genes and 2,291 cells
}
\keyword{datasets}
