% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_result.R
\name{plot_result}
\alias{plot_result}
\title{Plot the Fitted Model}
\usage{
plot_result(para, t, color, marginal, flag, y1, raw = y1, plot_title = NULL)
}
\arguments{
\item{para}{A numeric vector of the Estimated parameters of the predefined distribution,
length equals 6 or 7 depending on distribution(take 6 for Poisson or ZIP, 7 for NB and ZINB)}

\item{t}{A numeric vector of the input normalized pseudotime data of a given gene,
length equals the numbers of cells}

\item{color}{A string vector of length 4 to define plot color, default=\code{c('red', 'blue', 'orange', 'darkgreen')}}

\item{marginal}{A string of the distribution name. One of \code{Poisson}, \code{ZIP}, \code{NB} and \code{ZINB}.}

\item{flag}{A boolean variable, flag=T indicates Valley shape, flag=F indicates Hill shape}

\item{y1}{A vector of integers, representing the input expression counts of a given gene after transformation if having valley shape,
length equals the numbers of cells, y1=raw if Hill shape(flag=F)}

\item{raw}{A vector of integers, representing the input expression counts of a given gene,
length equals the numbers of cells, default=y1}

\item{plot_title}{A string for plot title, default=NULL}
}
\value{
A ggplot object
}
\description{
Plot the Fitted Model
}
\examples{
data("df")
y1<-df$Gene1
t<-df$Time
flag<-FALSE
para<-c(2.29,3.27,11.79,0.58,30.4,60.82)
marginal<-"ZIP"
color<-c('red', 'blue', 'orange', 'darkgreen')
plot_result(para, t, color, marginal, flag, y1)

}
\author{
Shiyu Ma, Lehan Zou
}
