% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runscGTM.R
\name{runscGTM}
\alias{runscGTM}
\title{Estimate Parameters in Single-cell Gene Expression Generalized Trend Model on a List of Genes}
\usage{
runscGTM(
  gene.vec,
  t,
  y1,
  gene_name = NULL,
  marginal = "ZIP",
  iter_num = 50,
  seed = 123,
  mc.cores = 2
)
}
\arguments{
\item{gene.vec}{A vector of integers, indicating targeted genes' index in the data.}

\item{t}{A numeric vector of the input normalized pseudotime data of a given gene,
length equals the numbers of cells}

\item{y1}{A vector of integers, representing the input expression counts of corresponding lists of genes,
number of rows equals the numbers of cells, number of columns equals the numbers of targeted genes}

\item{gene_name}{A vector of strings, indicates the genes' name used in the model}

\item{marginal}{A string of the distribution name. One of \code{Poisson}, \code{ZIP}, \code{NB}, \code{ZINB} and \code{Gaussian}.
default=\code{ZIP}}

\item{iter_num}{A single integer vector, indicates max number of iteration used in the PSO algorithm
that estimates model parameters}

\item{seed}{A numeric variable of the random seed, affecting parametric fitting of the marginal distribution.
default=123}

\item{mc.cores}{Number of cores used for computing.}
}
\value{
A tibble of summary results of genes
}
\description{
Estimate Parameters in Single-cell Gene Expression Generalized Trend Model on a List of Genes
}
\examples{
data("df")
res <- scGTM::runscGTM(gene.vec=1:3, t=df$Time, y1=df[,3:5],gene_name=colnames(df[,3:5]))
}
\author{
Shiyu Ma
}
