% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scGTM.R
\name{scGTM}
\alias{scGTM}
\title{Estimate Parameters in Single-cell Gene Expression Generalized Trend Model}
\usage{
scGTM(
  gene_index = 100,
  t = NULL,
  y1 = NULL,
  gene_name = NULL,
  marginal = "ZIP",
  iter_num = 50,
  seed = 123
)
}
\arguments{
\item{gene_index}{A single integer vector, indicates which gene to estimate in the model,
default=100}

\item{t}{A numeric vector of the input normalized pseudotime data of a given gene,
length equals the numbers of cells}

\item{y1}{A vector of integers, representing the input expression counts of a given gene,
length equals the numbers of cells}

\item{gene_name}{A single string vector, indicates the gene name used in the model}

\item{marginal}{A string of the distribution name. One of \code{Poisson}, \code{ZIP}, \code{NB} and \code{ZINB}.
default=\code{ZIP}}

\item{iter_num}{A single integer vector, indicates max number of iteration used in the PSO algorithm
that estimates model parameters}

\item{seed}{A numeric variable of the random seed, affecting parametric fitting of the marginal distribution.
default=123}
}
\value{
The log_likelihood cost, estimated parameters, and their confidence intervals of a gene
}
\description{
The model fits a gene's expression counts and its normalized pseudotime into one of the four marginal distributions.
This function estimates corresponding parameters and confidence intervals in the corresponding marginal distributions.
}
\examples{
y1<-floor(runif(100, min = 0, max = 20))
t<-runif(100, min = 0, max = 1)
marginal<-"ZIP"
scGTM(gene_index=1, t=t, y1=y1, marginal=marginal)

}
\author{
Shiyu Ma, Lehan Zou
}
